requirements.txt
# supabase

# migration from supabase to mongodb
pymongo
flask
python-dotenv
boto3
flask_cors
sentence-transformers
PyMuPDF
numpy
requests
llama-index
streamlit
langchain-cohere